chore: separate seed processing
