+++ shank-render/src/lib.rs
@@ -1,3 +1,6 @@
 pub mod consts;
+mod process_seeds;
 mod render_seeds;
+
+pub use process_seeds::*;
 pub use render_seeds::*;
